interfacing: move interfaces to specific files
