+++ server/src/interfaces/Weather.ts
@@ -1,0 +1,15 @@
+export enum WeatherType {
+    Default,
+    Now,
+    Hourly
+};
+
+export interface WeatherForecast {
+    forecast: Weather[];
+}
+
+export interface Weather {
+    temp: number;
+    weather: string;
+    dt: Date;
+}
